sesi 9 - Penambahan Fitur Login

+++ app/views/mahasiswa/index.php
@@ -178,23 +178,32 @@
 
                     <td>
 
-                        <a
-                            href="<?= BASEURL; ?>/mahasiswa/edit/<?= $mhs['id']; ?>"
-                            class="btn btn-warning btn-sm"
-                        >
-                            Edit
-                        </a>
+                        <?php if ($_SESSION['user']['role'] == 'admin') : ?>
 
-                        <a
-                            href="<?= BASEURL; ?>/mahasiswa/delete/<?= $mhs['id']; ?>"
-                            class="btn btn-danger btn-sm"
-                            onclick="return confirm('Yakin ingin menghapus data ini?')"
-                        >
-                            Delete
-                        </a>
+                            <a
+                                href="<?= BASEURL; ?>/mahasiswa/edit/<?= $mhs['id']; ?>"
+                                class="btn btn-warning btn-sm"
+                            >
+                                Edit
+                            </a>
+
+                            <a
+                                href="<?= BASEURL; ?>/mahasiswa/delete/<?= $mhs['id']; ?>"
+                                class="btn btn-danger btn-sm"
+                                onclick="return confirm('Yakin ingin menghapus data ini?')"
+                            >
+                                Delete
+                            </a>
+
+                        <?php else : ?>
+
+                            <span class="text-muted">
+                                Tidak ada akses
+                            </span>
+
+                        <?php endif; ?>
 
                     </td>
-
                 </tr>
 
             <?php endforeach; ?>
